requirements.txt
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
idna==2.8
pep8==1.7.1
pycodestyle==2.5.0
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.3
